+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 Unreleased
 ----------
+- Adjusted `test::WebSocketStream` alias to include
+  `tungstenite::MaybeTlsStream`
 - Switched from using `test-env-log` to `test-log`
 - Bumped minimum supported Rust version to `1.46`